pipeline can set time zone
Readings need the time zone to get DST right. OED already dealt with
this but assumed the provided reading had the offset. The pipeline can
now change a reading to have the time zone of the meter (or site if
none on meter).
Note this commit does not integrate into reading uploads.