remove core-js package
This package was listed by npm as extraneous so it was pruned.
This removed two babel packages and one other. This appears
to be left over from removing babel or some issue when that
work was merged. It now seems resolved and removes the
warning on install about old version of core-js. It is done here
rather than in an upcomig package update since rc-slider also
needed this.