Fixes bug/issues when msg too long
This seems to be a long standing bug. When the msg total
in processData is exceeded, a bad variable name caused the code
to die without warning. The name was fixed and now it works fine.

In addition, it now adds the warning on exceeding the allowed
length to the msg returned to user. It also makes minor wording
change.