modify group 3D to use group hourly view
This changed group 3D that used to used to sum the meter 3D results to directly
get the values from the new group hourly view. The idea is this is faster since it does
not sum over the meters.

As part of doing this, a function was created to determine the time for each point
and used in both the meter and group 3D case.

An index was added to mirror the one for hourly but added the new graphic unit.
A TODO was left since there are now two indices on the group hourly view and
it is unclear which are best/needed but adding this done substantially sped up the
3D data acquisition.

This changes the returned value when the request is inconsistent with 3D. For example,
the underlying meter frequency is too long. Previously it returned  -999 * # meters in
group but now it returns -999. This does not impact the code as it looks
for a negative value in this case. One test case was updated to reflect this.