try to fix merge issues
A number of files were messed up during the merge of timeVary. This
attempts to fix them but testing, etc. is not yet working. Unfortunately,
I could not figure out how to keep the original git attribution for most
lines so they how indicate I did it. I want to note that most were originally
done by @tfrendl and earlier commits show this.